requirements.txt
openai
pyyaml
pandas
